Clean: Removed some comments, deleted some redundant methods
