+++ src/core/tools.py
@@ -49,26 +49,6 @@ def timestamp_converter(timestamp: str, timestamp_type="short") -> str:
         return timestamp.strftime("%x %H:%M")
     return legible_timestamp
 
-def load_user_config() -> dict:
-    """Returns the user's NOMM configuration file data as a dictionary"""
-    user_config_path = os.path.join(GLib.get_user_data_dir (), 'nomm', 'user_config.yaml')
-    try:
-        data = load_yaml(user_config_path)
-    except:
-        print("Error: could not load user config.")
-        return None
-    return data
-
-def write_user_config(data: dict) -> dict:
-    """Writes to the user's NOMM configuration file"""
-    try: 
-        user_config_path = os.path.join(GLib.get_user_data_dir (), 'nomm', 'user_config.yaml')
-        write_yaml(data, user_config_path)
-    except:
-        print("Error: could not write to user config.")
-        return False
-    return True
-
 def translate_fuse_path(folder_info) -> str:
     folder_path = folder_info.get_path()
     if "run/user" in folder_path:
