+++ src/gui/application.py
@@ -9,8 +9,8 @@ gi.require_version('Notify', '0.7')
 
 from gi.repository import Adw, Gdk, GdkPixbuf, GLib, Gtk, Pango, Gio
 
-from core.config import update_user_config
-from core.tools import load_yaml, write_yaml, load_user_config, write_user_config, translate_fuse_path
+from core.user_config import update_user_config, load_user_config, write_user_config
+from core.tools import load_yaml, write_yaml, translate_fuse_path
 from core.scanner import get_steam_base_dir, scan_all_games
 from gui.app_views.library_view import LibraryView
 from gui.dashboard import GameDashboard
@@ -183,7 +183,6 @@ class Nomm(Adw.Application):
         self.temp_config["staging_path"] = selected_folder_path
         self.user_defined_paths.append(selected_folder_path)
         self.show_nexus_api_key_screen()
-
 
     def show_nexus_api_key_screen(self):
         self.remove_stack_child("api_key")
